commit jwt

+++ controllers/insertionController.js
@@ -6,19 +6,22 @@ const modifUrl = false;
 const dotenv = require('dotenv');
 dotenv.config();
 
-const getInsertions = async(req, res) => {
+const getInsertions = async (req, res) => {
 
     Insertion.find((err, result) => {
-        if(!err) {
+        if (!err) {
             res.status(200).send(result);
         } else {
-            return res.status(500).json({'error':'Erreur lors de la requête des dispositifs d\'insertion:'+err});
+            return res.status(500).json({ 'error': 'Erreur lors de la requête des dispositifs d\'insertion:' + err });
         }
-    }) 
+    })
 
 };
 
-const createInsertion = async(req, res) => {
+const createInsertion = async (req, res) => {
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
+
     const title = req.body.title;
     const min_age = req.body.min_age;
     const max_age = req.body.max_age;
@@ -33,15 +36,15 @@ const createInsertion = async(req, res) => {
     console.log(res);
     if (checkInsert(title, min_age, max_age, url, res)) return res;
     if (modifUrl) {
-        url = "https"+url.substring(4);
+        url = "https" + url.substring(4);
         modifUrl = false;
     }
-    Insertion.findOne({title: title}).then((insertion) => {
-        if(insertion) {
-            return res.status(400).json({'error':'Ce dispositif d\'insertion existe déjà'});
+    Insertion.findOne({ title: title }).then((insertion) => {
+        if (insertion) {
+            return res.status(400).json({ 'error': 'Ce dispositif d\'insertion existe déjà' });
         } else {
             const newInsertionRecord = new Insertion({
-                title : title,
+                title: title,
                 min_age: min_age,
                 max_age: max_age,
                 income: income,
@@ -55,33 +58,38 @@ const createInsertion = async(req, res) => {
             })
 
             newInsertionRecord.save((err, result) => {
-                if (!err)  {
-                    saveLog(jwt.verify(req.headers['authorization'], process.env.TOKEN_SECRET).sub, result.id, "Created insertion")
+                if (!err) {
+                    saveLog(userId, result.id, "Created insertion")
                     res.status(201).send(result);
                 } else {
-                    return res.status(400).json({'error':'Erreur survenue lors de la sauvegarde'});
+                    return res.status(400).json({ 'error': 'Erreur survenue lors de la sauvegarde' });
                 }
             })
         }
     })
 };
 
-const deleteInsertion = async(req, res) => {
-    id = req.body.id;
-    
+const deleteInsertion = async (req, res) => {
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
 
-    if(!ObjectId.isValid(id)) return res.status(400).json({'error':'L\'ID spécifié n\'existe  pas'});
-    Insertion.findByIdAndDelete(id,(err, result) => {
+    id = req.body.id;
+
+
+    if (!ObjectId.isValid(id)) return res.status(400).json({ 'error': 'L\'ID spécifié n\'existe  pas' });
+    Insertion.findByIdAndDelete(id, (err, result) => {
         if (!err) {
-            saveLog(jwt.verify(req.headers['authorization'], process.env.TOKEN_SECRET).sub, id, "Deleted insertion")
-            return res.status(200).json({'success':'Insertion supprimée avec succès'});
+            saveLog(userId, id, "Deleted insertion")
+            return res.status(200).json({ 'success': 'Insertion supprimée avec succès' });
         } else {
-            return res.status(500).json({'error':'Erreur lors de la suppression'});
+            return res.status(500).json({ 'error': 'Erreur lors de la suppression' });
         }
     })
 };
 
-const updateInsertion = async(req, res) => {
+const updateInsertion = async (req, res) => {
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
 
     const id = req.body.id;
     const title = req.body.title;
@@ -96,51 +104,53 @@ const updateInsertion = async(req, res) => {
     const goal = req.body.goal;
     const info = req.body.info;
 
-    if(checkInsert(title, min_age, max_age, url, res)) return res;
+    if (checkInsert(title, min_age, max_age, url, res)) return res;
 
-    if(!ObjectId.isValid(id)) return res.status(400).json({'error':'L\'ID spécifié n\'existe  pas'});
-    Insertion.findByIdAndUpdate(id,{$set: {
-        title : title,
-        min_age: min_age,
-        max_age: max_age,
-        income: income,
-        url: url,
-        audience: audience,
-        duration: duration,
-        funding: funding,
-        organisation: organisation,
-        goal: goal,
-        info: info
-    }},{ new: true }, function(err, result) {
-        if(err) {
-            return res.status(400).json({'error':'Echec de la modification'});
+    if (!ObjectId.isValid(id)) return res.status(400).json({ 'error': 'L\'ID spécifié n\'existe  pas' });
+    Insertion.findByIdAndUpdate(id, {
+        $set: {
+            title: title,
+            min_age: min_age,
+            max_age: max_age,
+            income: income,
+            url: url,
+            audience: audience,
+            duration: duration,
+            funding: funding,
+            organisation: organisation,
+            goal: goal,
+            info: info
+        }
+    }, { new: true }, function (err, result) {
+        if (err) {
+            return res.status(400).json({ 'error': 'Echec de la modification' });
         } else {
-            saveLog(jwt.verify(req.headers['authorization'], process.env.TOKEN_SECRET).sub, id, "Modified insertion");
-            return res.status(200).json({'success':'Modification réussie'});
+            saveLog(userId, id, "Modified insertion");
+            return res.status(200).json({ 'success': 'Modification réussie' });
         }
     })
 };
 
-const getInsertionByAge = async(req, res) => {
+const getInsertionByAge = async (req, res) => {
     const age = req.body.age;
 
-    Insertion.find({ $and: [{min_age: {$lte: age}}, { max_age: { $gte: age }}]}, (err, result) => {
-        if(!err) {
+    Insertion.find({ $and: [{ min_age: { $lte: age } }, { max_age: { $gte: age } }] }, (err, result) => {
+        if (!err) {
             res.status(200).send(result);
         } else {
-            return res.status(204).json({'error':'Aucun dispositif d\'insertion ne correspond a l\'âge donné'});
+            return res.status(204).json({ 'error': 'Aucun dispositif d\'insertion ne correspond a l\'âge donné' });
         }
     })
 };
 
-const getByOrganisation = async(req, res) => {
+const getByOrganisation = async (req, res) => {
     const organisation = req.body.organisation;
 
-    Insertion.find({organisation: organisation}, (err, result) => {
-        if(!err) {
+    Insertion.find({ organisation: organisation }, (err, result) => {
+        if (!err) {
             res.status(200).send(result);
         } else {
-            return res.status(204).json({'error':'Aucun dispositif d\'insertion ne correspond a l\'organisation donnée'})
+            return res.status(204).json({ 'error': 'Aucun dispositif d\'insertion ne correspond a l\'organisation donnée' })
         }
     })
 }
@@ -149,13 +159,13 @@ function checkInsert(title, min_age, max_age, url, res) {
     if (title == "" || title == null) {
         return res.status(400).json({ 'error': 'Veuillez renseigner le nom de l\'organisation' });
     }
-    if(min_age>max_age) {
-        return res.status(400).json({'error': 'L\'age maximum doit être supérieur au minimum'});
+    if (min_age > max_age) {
+        return res.status(400).json({ 'error': 'L\'age maximum doit être supérieur au minimum' });
     }
-    if(!(max_age>0 && min_age>0)) {
-        return res.status(400).json({'error':'Veuillez entrer un age valide'});
+    if (!(max_age > 0 && min_age > 0)) {
+        return res.status(400).json({ 'error': 'Veuillez entrer un age valide' });
     }
-    if(!(url.substring(0,5) == "https")) {
+    if (!(url.substring(0, 5) == "https")) {
         modifUrl = true;
         console.log("Url modifié");
     }
@@ -164,7 +174,7 @@ function checkInsert(title, min_age, max_age, url, res) {
 function saveLog(actorId, targetId, action) {
     const newLog = new Log({
         userId: actorId,
-        action: action+" id: "+targetId
+        action: action + " id: " + targetId
     })
     newLog.save((err, result) => {
         if (!err) {
@@ -175,4 +185,18 @@ function saveLog(actorId, targetId, action) {
     });
 }
 
-module.exports = { getInsertions, getInsertionByAge, getByOrganisation , createInsertion, deleteInsertion, updateInsertion };
+// fonction de vérification de la validité du token, renvoie null si erreur
+function verifyToken(req, res) {
+    try {
+        jwt.verify(req.headers['authorization'], process.env.TOKEN_SECRET, function (tokenErr, decoded) {
+            if (tokenErr) throw new Error(tokenErr);
+            req.auth = decoded;
+        })
+    } catch (e) {
+        res.status(403).json({ 'error': 'Token invalide ' + e });
+        return null;
+    }
+    return req.auth.sub;
+}
+
+module.exports = { getInsertions, getInsertionByAge, getByOrganisation, createInsertion, deleteInsertion, updateInsertion };
