+++ controllers/organisationController.js
@@ -2,16 +2,16 @@ const Orga = require("../models/organisationModel");
 const ObjectId = require('mongoose').Types.ObjectId;
 const ORGA_REGEX = /^[a-zA-Z0-9\-\s]+$/;
 
-const getOrgs = async(req, res) => {
+const getOrgs = async (req, res) => {
 
     let name = req.body.name;
     if (name) {
-        if(checkOrga(name, res)) return res;
-        Orga.findOne({name: name}).then((orga) => {
-            if(orga) {
-                return res.status(200).json({orga});
+        if (checkOrga(name, res)) return res;
+        Orga.findOne({ name: name }).then((orga) => {
+            if (orga) {
+                return res.status(200).json({ orga });
             } else {
-                return res.status(400).json({'error':'Aucune organisation ne correspond a ce nom'});
+                return res.status(400).json({ 'error': 'Aucune organisation ne correspond a ce nom' });
             }
         })
     } else {
@@ -25,7 +25,11 @@ const getOrgs = async(req, res) => {
     }
 };
 
-const createOrg = async(req, res) => {
+const createOrg = async (req, res) => {
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
+    if (adminCheck(userId) == false) return res;
+
 
     if (checkOrga(req.body.name, res)) return res;
 
@@ -42,33 +46,43 @@ const createOrg = async(req, res) => {
     })
 };
 
-const updateOrg = async(req, res) => {
+const updateOrg = async (req, res) => {
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
+    if (adminCheck(userId) == false) return res;
+
 
     let id = req.body.id;
     let name = req.body.name;
-    if(checkOrga(name, res)) return res;
+    if (checkOrga(name, res)) return res;
 
-    if(!ObjectId.isValid(id)) return res.status(400).json({'error':'L\'ID spécifié n\'existe  pas'});
-    Orga.findByIdAndUpdate(id,{$set: {
-        name: name,
-    }},{ new: true }, function(err, result) {
-        if(err) {
-            return res.status(500).json({'error':'Echec de la modification'});
+    if (!ObjectId.isValid(id)) return res.status(400).json({ 'error': 'L\'ID spécifié n\'existe  pas' });
+    Orga.findByIdAndUpdate(id, {
+        $set: {
+            name: name,
+        }
+    }, { new: true }, function (err, result) {
+        if (err) {
+            return res.status(500).json({ 'error': 'Echec de la modification' });
         } else {
-            return res.status(200).json({'success':'Modification réussie'});
+            return res.status(200).json({ 'success': 'Modification réussie' });
         }
     })
 };
 
-const deleteOrg = async(req, res) => {
+const deleteOrg = async (req, res) => {
+    const userId = verifyToken(req, res);
+    if (userId == null) return res;
+    if (adminCheck(userId) == false) return res;
+
 
     let id = req.body.id;
-    if(!ObjectId.isValid(id)) return res.status(400).json({'error':'L\'ID spécifié n\'existe  pas'});
-    Orga.findByIdAndDelete(id,(err, result) => {
+    if (!ObjectId.isValid(id)) return res.status(400).json({ 'error': 'L\'ID spécifié n\'existe  pas' });
+    Orga.findByIdAndDelete(id, (err, result) => {
         if (!err) {
-            return res.status(200).json({'success':'Organisation supprimée avec succès'});
+            return res.status(200).json({ 'success': 'Organisation supprimée avec succès' });
         } else {
-            return res.status(500).json({'error':'Erreur lors de la suppression'});
+            return res.status(500).json({ 'error': 'Erreur lors de la suppression' });
         }
     })
 };
@@ -82,6 +96,20 @@ function checkOrga(name, res) {
         return res.status(400).json({ 'error': 'Nom d\'organisation invalide (pas de caractères spéciaux)' });
     }
 };
+
+// fonction de vérification de la validité du token, renvoie null si erreur
+function verifyToken(req, res) {
+    try {
+        jwt.verify(req.headers['authorization'], process.env.TOKEN_SECRET, function (tokenErr, decoded) {
+            if (tokenErr) throw new Error(tokenErr);
+            req.auth = decoded;
+        })
+    } catch (e) {
+        res.status(403).json({ 'error': 'Token invalide ' + e });
+        return null;
+    }
+    return req.auth.sub;
+}
 
 //On exporte nos fonctions
 module.exports = { getOrgs, createOrg, updateOrg, deleteOrg };
